requirements.txt
discord.py[voice]
discord-components
pillow
